🐛 avoid double slashes in fetch URLs (#70)
By using proper URL operations instead of string concatenation.

Tested by linking into Rolling Rick configured with a trailing slash
base URL.

This fixes an issue where a base URL with a trailing slash resulting in
a double slash in the fetch URL would run into CORS issues because while
fun-fair does redirect the double slash to single slash, it does so
without CORS Allow.

I'll publish a client release afterwards, but it is not urgent to update
games to the new version. The base URLs are in our control so we can
just avoid the issue by not using trailins slash base URLs.