📝 extend client documentation (#143)
Expanded a bit on API functions that you'd usually use, and when you'd use them.

But from talking to @makando95, it sounded to me more like what's
actually considered missing here is some context on how games must
fundamentally operate, so that they can be used on our platform.
This doesn't technically belong in the client docs specifically, but
doesn't hurt as extra context.